Add login keypress event and refactor UI layout
This commit includes the addition of a keypress (Enter key) event in the login form to improve user experience. The layout of the app has also been modified, making it dependent on the login state and there are changes in the AppBar rendering as well. Several miscellaneous refactors improve code readability and efficiency.